requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
streamlit
PyPDF2
google.generativeai
langchain_openai 
langchain_core
langchain_community
langserve
langchainhub
sentence_transformers
langchain-objectbox
sse_starlette
